REF: ChallengesController Class with Dependency Injection -The controllers is now a class to allow us to use dependency injection to pass the repository to the class constructor. -The controller cannot instantiate repositories to pass it to the service, as it is not part of the infrastructure layer, so it should receive it through injection. -Refactored the method names to follow the verbs convention for controllers (i.e. index, show, create, update, etc). -The utils filtering is used by the controller because filtering pagination and sorting is part of the application layer. They are not business logic. -ChallengesController.js was moved from the controllers folder into the challenges module.
